Implement StreamSource interface and refactor stream handling in sources package
